# Much faster implementation of FileList, for big egg_info speedups