Refs #249 and #450 

This is not a complete PR, but request for the approach to be confirmed before I do more work.

There are some notes:

# Code

The approach, as outlined before, is to reduce `findall` to a no-op, so the directory tree is not scanned, and have `include_pattern` (which handles all 'include' type directives) execute an `os.walk` directly, but trimming the search tree according to the different types of rules.

It currently doesn't handle all cases correctly (specific `graft` doesn't work correctly), but is a proof of concept. If the approach is accepted, I would add more tests (since we can no longer rely on distutils have tests since we are overriding the implementation) and fix this bug. All the tests in setuptools pass with the patch.

Some testing on Windows will be necessary, the directory path separator thing gets quite involved, I don't have a Windows box I could easily use.

Performance tests are hard - we really want to test that other parts of the filesystem are *not* being scanned, which is hard to do.

## Removal of `_add_egg_info` function, called from `add_defaults`

I believe this is correct. The behaviour of `add_defaults` seems to be tested by `test_egg_base_installed_egg_info` and `test_install_requires_with_markers` which both fail if you break it, along with a few other tests that fail.

AFAICS, the code is only working around the fact that `include_pattern` queries the `allfiles` attribute and not the filesystem, and is therefore out of date with respect to files that have been created by setuptools itself. Since we've changed the behaviour of include_pattern, I don't think this is necessary, and the tests seem to verify this.


# Metrics

Obviously this depends on how many files you've got in your working folder, and on whether the directory listings are in the OS cache.

I used this project as a not-too-extreme example, and because I actually work on this project:

https://github.com/django-money/django-money

It uses tox to test against about 80 environments, which are stored as virtualenvs in .tox

## Without patch

From a 'cold' state, but with virtualenvs created from previous usage, I ran the tests on a single environment:

    tox -e django18-py27

This took 647 seconds (10 minutes), of which just 2.8 seconds were the actual tests i.e. 645 seconds of overhead from setuptools and tox.

Doing it a second time, this reduced to 37 seconds of overhead, due to OS caching of filesystem.

## With this patch

With the patch, the setuptools + tox overhead of the same command reduced to just 2 seconds, whether from warm or cold i.e. a speed up of 18 times to 320 times (and more importantly, down to "a few seconds", an acceptable time to wait for your tests to start running). This would make a BIG difference to workflow on quite a few projects I use.

To reproduce this speedup, however, you will need to install the dev version of setuptools in both the outer project (in which tox is installed) and in the virtualenvs it created.

There are obviously cases and projects where this strategy will be less efficient than the previous one, because it requires multiple `os.walk` walks over the file system, instead of just one, but for normal projects, and for any that contain large numbers of non source files, it is likely to be a very big
win.

## Notes

There will be no speedups if MANIFEST.in contains a global-include directive, and in fact probably severe slow-downs if you have lots of files or lots of global-include directives. This is because you cannot trim the search tree at all for this directive - a global-include includes all subdirectories, including hidden ones such as ".tox", ".my_IDE_like_to_put_stuff_here" etc. This means that in general the global-include directive should be avoided, because most tools expect to be able to put all kinds of things in hidden folders and have them ignored. This warning should probably be submitted to distutils as a documentation patch.