@jaraco 2016-06-24
The main reluctance I have with this approach is it increases the reliance on `MANIFEST.in`, which is a duplicative, redundant functionality with source-code control systems (as [I mentioned in this comment](https://github.com/pypa/setuptools/issues/450#issuecomment-202921921)). The SCM systems keep a manifest of files in the project and have mechanisms to ignore whole trees (e.g. `.gitignore`). Furthermore, these systems provide a robust, layered mechanism for ignoring files and trees at the system or user level in addition to the project level. The more we can leverage this functionality, which is almost universally adopted, the more we can rely on these advances. I imagine developing from non-SCM tarballs is rare and that issues like #249 can be disregarded in those cases.

I'd really prefer a solution that relies on SCM file finders and limit the reliance on the (soft-) deprecated MANIFEST.in. If a SCM file finder solution is nonviable, I'd at least like to have robustly determined such before adopting another approach.