@spookylukey 2016-06-25
In what way is this an increased reliance on MANIFEST.in? The situation is exactly the same as before, my patch is simply a change to the implementation of the MANIFEST.in processing to be (typically) much faster than the one inherited from distutils.

It would be nice if this wasn't necessary, but at the moment, using an SCM file finder will not improve the performance problems with egg_info, because `FileList.findall()` runs as well as the SCM file finder. AFAICS, this behaviour can't change, because:

1) Sometimes you don't want to package everything under source control
2) Sometimes you want to add something not under source control (e.g. a file generated from one under SCM).

Number 1 is explicitly talked about in the setuptools-git docs - https://pypi.python.org/pypi/setuptools-git

Is there actually a plan to change this and a way forward to solve these?

With this in mind, I can't see how it is possible to rely on an SCM file finder to 'provide' optimisation. The only thing I can think of is if we use '.gitignore' etc. files to avoid searching into directories. However, this is full of problems:

1) `.gitignore` files will probably list generated files/directories, but we might actually want these in distributed tarballs.
2) It could easily get complicated by SCM 'global' ignore files (e.g. I have `.tox` in my global `.gitignore`, not local)
3) It would require a lot of additional functionality in setuptools, interfacing with different SCMs and all the ways they might specify files to ignore.

It also seems a very strange dependency - why do we need an SCM tool to tell us not to bother looking in certain directories, when there wasn't a reason to look in them in the first place? With hindsight, "first get a list all the files in all the folders" was simply not a very good strategy for implementing the MANIFEST.in language, and the solution is simply "don't do that", which is what this patch implements.