@jaraco 2016-07-24
> In what way is this an increased reliance on MANIFEST.in? 

If someone wants to gain the performance benefits of this patch, they will be required to create a MANIFEST.in, even if the project is relying on SCM info, and adds confounding factors to its purpose.

What I'd really like to see is a logic flow where file discovery techniques are clearly described and selected... something like a Discovery class with three implementations, one for Manifest.in, another for SCM, and another for the most simple heuristic (traversal including *.py files). And exactly one of those is chosen.

What worries me most is this logic feels interwoven with an already complicated implementation. Every time I look at it, I feel like I don't fully grasp all the implications. I worry the added complication will lead to more problems with the design down the road, and it won't be obvious where this functionality exists or why.

But I'm probably just overcautious. Do feel free to continue with the effort. We can roll it out and see what reception it gets.