@timheap 2016-08-15
Many projects I work on contain files that are tracked in git that should not be packaged (such as docs, tests, etc), and files that are not tracked but should be packaged (such as compiled JS, CSS assets). Integrating with git or any other SCM would be counter productive for the projects I work on, and I suspect many others as well.

While the overhead in the projects I work on is not quite as extreme as for @spookylukey, it is still noticeable. A fix would be nice. I've also submitted a patch for pip, and another for setuptools (#733) for similar speed improvements. I was very happy to see a fix for this issue already submitted!

@jaraco I understand that you desire a more flexible solution to this problem, with multiple ways of specifying files to include. I've also struggled with the limitations of the MANIFEST file format. This PR solves a problem that people are experiencing right now though. Merging this PR would not prevent you from implementing your Discovery class in the future, and when the Discovery class is implemented this solution could be made more generic so more classes could reuse it.

Is there anything I can do to help move this PR along?