@jaraco 2016-08-15
I believe this PR is a work in progress. My last message indicated that we should roll it out, so I'm waiting on a follow-up to say that it's ready to go.

@timheap, you can help by reconciling which of these two patches are preferable (or justifying how they're both desirable).