@timheap 2016-08-15
@jaraco This patch and #733 address different performance problems in different areas, so they are both desirable. They affect unrelated sections of setuptools, the only similarity is the performance fix.

I'll have a look to see if I can help finish this PR as well. I installed the combination of this PR and mine and it made setuptools very speedy!