@spookylukey 2016-08-16
@timheap thanks for the offer of help.

I'm unlikely to have time to look more at this in coming months. In terms of implementation, I don't think there was a huge amount more to do on this PR, although there is at least one TODO in there.

However, more tests would be nice. I'm afraid I can't remember how much of this code had test coverage. I think it probably needs more testing on other platforms.