@spookylukey 2016-08-16
@jaraco thanks for your continued help/patience with this PR, by the way!