@jaraco 2016-10-14
I'm going to close this PR, but to be sure, I'm welcoming further work on it. Just open a new one when you have more code for consideration.